+++ js/src/admin/components/Settings.js
@@ -18,6 +18,11 @@ export default class Settings extends ExtensionPage {
 			name: 'Flarum Blog',
 			url: 'https://flarum.org/extension/v17development/flarum-blog',
 		};
+		const repostExtension = {
+			id: 'shebaoting-repost',
+			name: 'Repost',
+			url: 'https://flarum.org/extension/shebaoting/repost',
+		}
 
 		const warningIcon = icon("fas fa-exclamation-triangle");
 
@@ -202,6 +207,29 @@ export default class Settings extends ExtensionPage {
 								label: app.translator.trans("walsgit_discussion_cards.admin.settings.general.useBlogSummary_label"),
 								help: app.translator.trans("walsgit_discussion_cards.admin.settings.general.useBlogSummary_help"),
 								disabled: isExtensionActive(blogExtension.id) ? false : true,
+							})}
+							<hr></hr>
+							{/* shebaoting/repost */}
+							<h4>
+								{app.translator.trans("walsgit_discussion_cards.admin.settings.general.repostExtension_title_start")}
+								<a href={repostExtension.url} target="_blank" rel="noopener noreferrer">{repostExtension.name}</a>
+								{app.translator.trans("walsgit_discussion_cards.admin.settings.general.repostExtension_title_end")}
+							</h4>
+							<div className="helpText">
+								<em>
+									{!isExtensionInstalled(repostExtension.id) 
+									? + app.translator.trans("walsgit_discussion_cards.admin.settings.general.repostExtension_notInstalled", {icon: warningIcon}) : 
+									!isExtensionActive(repostExtension.id) 
+									? app.translator.trans("walsgit_discussion_cards.admin.settings.general.repostExtension_notActivated", {icon: warningIcon}) :
+									''}
+								</em>
+							</div>
+							{this.buildSettingComponent({
+								type: "switch",
+								setting: "walsgit_discussion_cards_allowRepostLinks",
+								label: app.translator.trans("walsgit_discussion_cards.admin.settings.general.allowRepostLinks_label"),
+								help: app.translator.trans("walsgit_discussion_cards.admin.settings.general.allowRepostLinks_help"),
+								disabled: isExtensionActive(repostExtension.id) ? false : true,
 							})}
 						</div>
 						{this.submitButton()}