+++ CHANGELOG.md
@@ -1,10 +1,11 @@
 # Changelog
 
-## [1.2.0] - 2024-12-11
+## [1.2.0] - 2024-12-12
 ### Added
 - NEW option to show the number of replies on the right side of a card's title (like the default Flarum design) instead of on the image.
 - NEW option to show the last post informations on the cards (like the default Flarum design): username & date of last post.
-- New migration file for the aforementioned new options.
+- NEW FEATURE: support for `shebaoting/repost` extension.
+- New migration file for the aforementioned new options & features.
 
 ### Changed
 - The way read cards are marked (removed B&W filter): now read cards' title and text have a lighter font weight (like the default Flarum discussion list behavior)